add destroy in watch

+++ src/module/directives/ueTextarea/ueTextarea.controller.js
@@ -258,7 +258,9 @@
         }, true);
 
         this.$onDestroy = function() {
-            destroyWatchEntityLoaded();
+            if (angular.isFunction(destroyWatchEntityLoaded)) {
+                destroyWatchEntityLoaded();
+            }
             destroyEntityLoaded();
             destroyErrorField();
             destroyWatchFieldValue();
